commit before push

+++ src/test/java/ro/msg/learning/shop/service/integrationTest/SupplierServiceTest.java
@@ -4,18 +4,13 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 import ro.msg.learning.shop.dto.SupplierDTO;
-import ro.msg.learning.shop.entities.Supplier;
 import ro.msg.learning.shop.repositories.SupplierRepository;
 import ro.msg.learning.shop.services.SupplierService;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
@@ -38,12 +33,12 @@ public class SupplierServiceTest {
         supplierService.create(sup2);
         supplierService.create(sup3);
 
-        int supplierNbService = (int) supplierService.listAll().size();
-        int supplierNbRepo = (int) supplierRepository.findAll().size();
-        int idOfSupplierOne=supplierRepository.findByName("NameSupplier1").orElseThrow().getId();
+        int supplierNbService = supplierService.listAll().size();
+        int supplierNbRepo = supplierRepository.findAll().size();
+        int idOfSupplierOne = supplierRepository.findByName("NameSupplier1").orElseThrow().getId();
 
-        Assert.assertEquals(3,supplierNbRepo);
-        Assert.assertEquals(supplierNbService,supplierNbRepo);
-        Assert.assertEquals(1,idOfSupplierOne);
+        Assert.assertEquals(3, supplierNbRepo);
+        Assert.assertEquals(supplierNbService, supplierNbRepo);
+        Assert.assertEquals(1, idOfSupplierOne);
     }
 }
